+++ src/lib.rs
@@ -78,8 +78,9 @@ pub mod types;
 pub const CODEC_ID_STR: &str = "tga";
 
 pub use decoder::{
-    parse_tga, parse_tga_colour_correction_table, parse_tga_developer_area,
-    parse_tga_extension_area, parse_tga_footer, parse_tga_postage_stamp, parse_tga_scan_line_table,
+    parse_tga, parse_tga_attributes_type, parse_tga_colour_correction_table,
+    parse_tga_developer_area, parse_tga_extension_area, parse_tga_footer, parse_tga_postage_stamp,
+    parse_tga_scan_line_table,
 };
 pub use encoder::{
     encode_tga_grayscale, encode_tga_grayscale_rle, encode_tga_palette, encode_tga_palette_rle,